Update CredentialLeakDialogBridge to use incognito profile when called incognito.
This CL updates the profile passing to HelpAndFeedback activity from
CredentialLeakDialogBridge. Currently, it is always regular profile,
however it is not safe because of possible data leakage from incognito
to regular profile.

This CL updates CredentialLeakDialogBridge to receive the current
profile (i.e., regular or off-the-record profile) of current tab and
pass it to HelpAndFeedback activity.

Bug: 1041781
Change-Id: Ia36d7d604d41f1dff72ed7acd4ac6dfcdafdb3bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063020
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Side YILMAZ <sideyilmaz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742699}