Change NetworkContextParams::use_builtin_cert_verifier from a bool to a tri-state.
The 3 values are kDefault, kBuiltin, kSystem. When the value is
kDefault, the NetworkContext will check
net::features::kCertVerifierBuiltinFeature for which verifier to use.

This means that services_unittests and other build targets which use the
network service but don't have explicit handling for the
use_builtin_cert_verifier setting will now use the expected default
implementation of the verifier instead of always using the system
verifier.

Bug: 1049869
Change-Id: I74d20c2bb153e34074dc35269927b582eb5c61a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2058004
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Commit-Queue: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742703}