Fixup WebSchedulerTrackedFeature Enum
Recently updating the WebSchedulerTrackedFeature histogram enum it was
noticed that the current labels are out of sync. This change adds a new
presubmit check to help warn committers to keep the value up to date.
In addition the current labels have been updated to reflect the current
in-code values being reported.

Change-Id: I8e3fb6dd452454f4efa0aecf65348bcce4517c6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063359
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Brandon Maslen <brandm@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#742822}