Move PermissionRequestManager to //components/permissions
This required moving a few other classes:
- NotificationPermissionUiSelector
- PermissionPrompt
- MockPermission*

The GetDisplayNameOrOrigin() method which used to be part of
PermissionPrompt::Delegate was moved to PermissionPromptBubbleView
since that is the only place it's used.

A set of unit tests was left in chrome which tests chrome-specific
behavior of the PermissionRequestManager. The rest of the tests were
moved into the component.

Bug: 1025609
Change-Id: Ib4420a4322adeccfa0fe42c6197a9111d9a100a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2057965
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742829}