Remove dead code for optimizations of empty contents and background
This is a follow-up for crrev.com/c/2031429 which disabled the
optimization of descendant background paint phase, and
crrev.com/c/2042262 which disabled the optimization of paint invalidation
for empty contents. The CLs caused no regression based on UMA results, and
acceptable regression for several perf tests, so we decide to move
forward to remove these regressions and implement the simplest solution
of hit testing of transparent backgrounds in CompositeAfterPaint.

Bug: 1049440
Change-Id: Ib21a02d893bffe724c0b2db0bb6e46953e8d0a48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2063251
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742879}