device/bluetooth: Add bluetooth suspend notifier flag
ChromeOS BlueZ is adding better support for system suspend and it can be
enabled by setting the UseKernelSuspendNotifier property. Add a
BluetoothKernelSuspendNotifier feature flag, expose it in chrome://flags
and use it in device/bluetooth to enable or disable the feature when the
adapter becomes available.

BUG=b:147884615
TEST=Built for Sarien, deploy_chrome to physical device, toggled flag in
     guest mode while running dbus-monitor to confirm property changes.
R=hansberry,mcchou

Change-Id: I6cdc874e2adaa7fb59435d1768fb29bffe4e424a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029258
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742908}