[system-web-apps] Add an interface to add frame-ancestors in CSP.
Add methods in WebUIImpl to add frame ancestors in the CSP which will
then be added to the response header.

Currently how WebUIs are embedded is by disabling an DenyXFrameOption()
in the headers that allows any page to embed a WebUI. We want to use
FrameAncestors in the CSP headers now and whitelist which pages we can
embed a WebUI in. For now, we will only use frame ancestors if
DenyXFrameOptions() is true. In the future, WebUIs that disable
DenyXFrameOptions will be refactored to use frame ancestors instead.

We have also refactored a function in WebUI browsertests where instead
of passing multiple parameters which are optional, we will now pass in
a ContentSecurityPolicy object.

This cl includes a fix to a bug where previously in the
WebUIURLLoaderFactory, the resource response's content security policy
was not set.

Bug: 1023700
Change-Id: I2612994c628c01033f7a30d784db66b12796c0b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029529
Commit-Queue: Jessica Huang <jessing@google.com>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: calamity <calamity@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742911}