Do not send input events if the event is located at the caption
Currently, if no window is found for a located event for a
captured window, it simply falls back to the main surface. This
means that click events on the frame (title bar) will be sent
to the underlying system through wayland, but that's unnecessary
at all.

This CL checks the located event's location through hit-test
component.

Bug: b/149517682
Test: exo_unittests
Change-Id: I2c314fca3aadea9f81475e7223166b0087b1e0ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062582
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742927}