Adjust PIP bounds before saving snap fraction.
This is a follow-up CL of crrev.com/c/2040505, which introduced a
regression where, if the client sends a bounds that is not along
any of the edges of the PIP movement area, a wrong snap fraction
is saved to the window.

To avoid this, this CL ensures to snap the bounds to one of the
edges.

BUG=b:204050
TEST=atest android.server.am.ActivityManagerPinnedStackTests
TEST=#testEnterPictureInPictureSavePosition on kevin

Change-Id: I9d617790fb31c1a9c9bf33cf21ff67d2ae923e02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2058934
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Cr-Commit-Position: refs/heads/master@{#742991}