ozone/wayland: create normal surface if no parent is found
Some tests may want to create a popup window without a parent. In case
of Wayland, it requires to pass a parent window's widget. If there is no
parent, and we are unable to find a current focused window (by cursor or
by keyboard), what we can do is to create a normal surface.

Bug: 578890
Change-Id: Iacbddd0cb85f62ea1f59fbf3efc6a160a3047fbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062976
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743000}