Update SadTab to use incognito profile when called incognito.
This CL updates the profile passing to HelpAndFeedback activity from
SadTab. Currently, it is always regular profile, however
it is not safe because of possible data leakage from incognito to
regular profile.

This CL updates SadTab to receive the current profile (i.e., regular or
off-the-record profile) of current tab and pass it to HelpAndFeedback
activity.

Bug: 1041781
Change-Id: I6c5b85187f627348e2dadbba8eaa3a18d02ddb94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062972
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Side YILMAZ <sideyilmaz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743030}