[Android] Photo Picker: Log the query used to enumerate media files.
When troubleshooting why no media files appear in the PhotoPicker
it is important to know what query was issued so that the file system
can be inspected (to see if any files should have matched).

This is only logged once (when a PhotoPicker dialog is opened).

Bug: 656015
Change-Id: I833665827f598bb1c40cbfeec2cc44f6c8c4b7d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2066537
Auto-Submit: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743088}