Clip focus ring points to screen
The screen reader's focus rings have outsets around the bounding box
from the semantics tree info.  However, when the bounding box is for
a full screen item (like the background image, for example) the
focus ring becomes invisible because it renders outside the screen
area.  This CL clips the points to the screen's dimensions.

Bug: 1052121
Test: Local build display assistant with screen reader enabled
Change-Id: Icf4ad68f59c23e503bc3baf9ec7763e175095853
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055668
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Commit-Queue: Randy Rossi <rmrossi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743110}