Make NetworkIsolationKey and site-for-cookies agree in some tests.
NetworkIsolationKey and site-for-cookies are related, in such a way as
that we shouldn't be sending same-site-cookies for requests when the
NIK doesn't indicate the top-frame origin (and frame origin, if
present) are for that site. We want to add DCHECKs that this is always
the case.  This CL fixes a couple NetworkContext unit tests where
the two values don't agree.

Bug: 1052148
Change-Id: Ibd7118310ac01b19f60001a1296f3e76af091c70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055426
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743114}