Pass real user-id to LibAssistant
Previously "0" was always passed as the user-id to LibAssistant.
This now causes issues when trying to introduce marble,
as marble actually relies on the user-id being a real Gaia-id.

Another change I've made here is introduce a |ScopedAshSessionObserver|
class which is used inside the Assistant |Service|.
This simplifies the house-keeping inside the |Service| class.

Note: This requires crrev.com/i/2556323 to land first
(and be uprevved into Chromium).

Bug: b/143984405
Change-Id: I28b6801a598de1b7f76c19af272d3fc18efa47e9
Tests: New chromeos_unittests that are covered by the filter "*Assistant*"
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2055428
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743125}