Fix SharedWorker URLLoaderFactory NetworkIsolationKey.
SharedWorkerServiceImpl was using a SiteForCookies based on the URL of
the worker being loaded, but a NetworkIsolationKey based on the URL of
the RenderFrameHost. In practice, these are always the same...except for
the case of extensions loading a SharedWorker of some site. This CL
fixes the issue.

Bug: 1052148
Change-Id: I53d37354defdcca5fe0ea3f1412ee7a638016a5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2057525
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743140}