Add a histogram entry for same-site but cross-origin downloads
This CL adds a new bucket to Download.InitiatedByWindowOpener to
indicate that a download is initiated by a same-site but
cross-origin opener of that tab. E.g., if a tab
subdomain1.example.com opens another tab subdomain2.example.com and
then navigates it to a download, this CL will now record the download
as kSameSite instead of kCrossOrigin.

This is added because latest number from Canary for the kCrossOrigin
bucket is unexpectedly large. I suspect this is caused by sites serving
their downloads on subdomains. This would previously go to the
kCrossOrigin bucket but it now goes under kSameSite bucket.

Bug: 121259
Change-Id: Idf142448b052a342df057982cfb79b0669fbd229
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2065252
Auto-Submit: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743183}