Allow BottomSheetContent to intercept back button
This patch adds the optional ability for sheet contents to intercept
the back button. This is useful in cases where the bottom sheet has an
unusually large peek state and back would be used to dismiss the sheet
rather than inadvertently navigate back on the underlying page.

Bug: 1049942
Change-Id: I3d1ff85f122a96743ea33a71b784f49e36c6e9ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2062748
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Donn Denman <donnd@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#743239}